Store `currentChain` in `ChainContext`
