+++ package/src/components/KitProvider/ChainContext.tsx
@@ -1,15 +1,24 @@
-import React, { createContext, useContext, useEffect, useMemo } from 'react';
+import React, {
+  createContext,
+  useContext,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
 import { Chain } from 'viem';
-import { useWalletClient } from './KitProvider';
+import { usePublicClient, useWalletClient } from './KitProvider';
+import { getChain } from '../../utils/utils';
 
 export interface ChainContextProps {
   supportedChains: Chain[];
   initialChain?: Chain | number;
+  currentChain?: Chain;
 }
 
 const ChainContext = createContext<ChainContextProps>({
   supportedChains: [],
   initialChain: undefined,
+  currentChain: undefined,
 });
 
 interface ChainContextProviderProps extends ChainContextProps {
@@ -22,6 +31,17 @@ export const ChainContextProvider = ({
   children,
 }: ChainContextProviderProps) => {
   const walletClient = useWalletClient();
+  const [currentChain, setCurrentChain] = useState<Chain | undefined>();
+  const publicClient = usePublicClient();
+
+  useEffect(() => {
+    const fetchChain = async () => {
+      const block = await publicClient?.[supportedChains[0].name]?.getChainId();
+      setCurrentChain(getChain(block || 1));
+      return;
+    };
+    void fetchChain();
+  }, [publicClient]);
 
   useEffect(() => {
     const switchChain = async () => {
@@ -42,6 +62,7 @@ export const ChainContextProvider = ({
           supportedChains,
           initialChain:
             typeof initialChain === 'number' ? initialChain : initialChain?.id,
+          currentChain,
         }),
         [supportedChains, initialChain]
       )}
@@ -55,3 +76,5 @@ export const useSupportedChains = () =>
   useContext(ChainContext).supportedChains;
 
 export const useInitialChainId = () => useContext(ChainContext).initialChain;
+
+export const useCurrentChain = () => useContext(ChainContext).currentChain;
